requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
sqlite3
pytest==7.4.3 